Added ModalityLayout and made sure PopupLayout is inheriting it
