+++ src/DIPS.Xamarin.UI/Controls/Modality/ModalityLayout.xaml
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<ContentView xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="DIPS.Xamarin.UI.Controls.Modality.ModalityLayout">
+    <RelativeLayout x:Name="relativeLayout"/>
+</ContentView>
