+++ src/DIPS.Xamarin.UI/Controls/Popup/PopupLayout.xaml
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
-<ContentView xmlns="http://xamarin.com/schemas/2014/forms"
+<Modality:ModalityLayout xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:Popup="http://dips.xamarin.ui.com"
+             xmlns:Modality="clr-namespace:DIPS.Xamarin.UI.Controls.Modality;assembly=DIPS.Xamarin.UI"
              x:Class="DIPS.Xamarin.UI.Controls.Popup.PopupLayout">
-    <RelativeLayout x:Name="relativeLayout"/>
-</ContentView>
+</Modality:ModalityLayout>
