+++ src/DIPS.Xamarin.UI/DIPS.Xamarin.UI.csproj
@@ -24,6 +24,12 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Update="Controls\Modality\ModalityLayout.xaml.cs">
+      <DependentUpon>ModalityLayout.xaml</DependentUpon>
+    </Compile>
+    <Compile Update="Controls\Popup\PopupLayout.xaml.cs">
+      <DependentUpon>PopupLayout.xaml</DependentUpon>
+    </Compile>
     <Compile Update="Controls\RadioButtonGroup\RadioButtonGroup.xaml.cs">
       <DependentUpon>RadioButtonGroup.xaml</DependentUpon>
     </Compile>
